Simplify (and fix) todo finish/reopen logic
After refactoring the javascript broke, because we no longer return the
nested json.
Also remove the javascript used for datetime string, and remove the need
for parsing that string; instead simply adding the datetime.now() on the
server side, when the is_finished value is True

Minor clean up on change_title API, where the nested json was still
present.