Simplify Dockerfile
Instead of copy/pasting some configuration, I dug deeper and understood
better what was actually required and could (as and added bonus) even
reduce the size of the resulting image. Before this change the build
image was 88.6MB and now it is 77.8MB!
This is mostly due to not caching the packages installed via pip.